turning cat a lot into separate feature
